feat(*): re-organize home and login components, give them separate templates and styles for demonstration
