+++ src/app/components/home/home.component.html
@@ -1,0 +1,4 @@
+<div [formGroup]="messageForm">
+  <h1>{{messageForm.controls['messageText'].value}}</h1>
+  <input type="text" formControlName="messageText"/>
+</div>
