+++ src/app/components/home/home.component.scss
@@ -1,0 +1,5 @@
+div {
+  & > input {
+    color: blue;
+  }
+}
