+++ src/app/components/login/login.component.html
@@ -1,0 +1,4 @@
+<div>
+  Authenticate
+  <button *ngIf='!authenticated' (click)='authenticate()'>Authenticate with Github</button>
+</div>
