+++ src/app/components/login/login.component.scss
@@ -1,0 +1,5 @@
+div {
+  & > button {
+    color: blue;
+  }
+}
